Convert altitudes and magnitudes to floats
Converted the altitude and magnitude values to floats. This is an
attempt to speed up computations on the gpu. Initial testing on my
machine (NVIDIA MX130) shows an average kernel speedup of the iterate
particle kernel from ~48ms to ~20ms. Setting up the simulation does
appear to take longer than before by about 1 second, this appears to
most likely come from the getLambdaAtS() function on line 83 of
dipoleB.cu. The increased error of using floats over doubles still needs
to be analyzed.