Setup MultiGPU Efield
This commit sets up the EField to run on multiple gpus. The simulation
class will keep an array to the pointer pointing to the EField The
desired device number is used to index into the array. Care was used
to still work with for a single gpu setups in addition to the running on
just the CPU. Compiler macros are used to access cuda functions in
the simulation.cpp file.