Setup Satelite MultiGPU
Setup the Satelite Class vectors to work for multiple gpu processing.
The vector length is now effectively the number of satelites * device
count. The vector is index as follows for 3 satelites and 2 devices:
{Sat 1 Dev 0, Sat 1 Dev 1, Sat 2 Dev 0, Sat 2 Dev 1}. The iterate
simulation function has already been setup to operate this way, but it
needs to be tested if it actually works.