Update Satelite Class for MultiGPU
Updated the Satelite class for mutligpu usage. It now works by having
arrays for the number of devices/particles for each device. As well
as arrays for the necessary data pointers for each device. When copying
the data back to the host, each device's data is copied to effectively
its own vector. The vectors are then transposed into a single data_m
vector. I have validated that the code runs on my single GPU machine.
I am not able to test for multi=gpu. If there is an error, it is likely
in copying data back as I was not 100% on how the data_m array was
arranged.

NOTE: The case where the satelite data could be loaded from a file was
not handled as the extents of the projects don't cover it. If desired
later, it is likely possible in a similiar manner as copying data from
device to host.